fix dat lich
trả về kq không đợi
